Added additional functionality to play now page
Includes:
-randomizing order of questions
-checking for correct/incorrect answer on all types of questions
-next question functionality
-keeping track of the position in the question array to be able to tell
the user when there are no more questions for the unit

All of these are implemented in html and css just so that I could see
that they worked. It’s not pretty by any means. Generally the question
stuff should stay the same, but otherwise the formatting of the page
can be played with a bunch